labgeneric: bootmode call python function
with the new entry in BOOTMODE_BOARDNAME section:

modes = {"name":"bootmode:usb", "func":"<name of function which gets called"}]

example:
modes = [{"name":"bootmode:usb", "func":"cxg3_setbootmode_usb"}, {"name":"bootmode:emmc", "func":"cxg3_setbootmode_emmc"} ]

it is possible to define a function which gets called when
this flag is passed to tbot.

For the above example, if tbot flag bootmode:usb is passed
the function cxg3_setbootmode_usb is called. You have to define
this function in a new file:

tbotconfig/labcallbacks.py

the functions has only one argument, the lab host.

This is a hack, better to add an implementation, but useful
for fast trying out stuff

Signed-off-by: Heiko Schocher <hs@denx.de>