ubootbuild: rework binaries upload
store the needed binaries in a subdirectory on lab host
in tftp subdirectory "binariessubpath", instead to store
them in tbottest subdirectory in U-Boot source (where we
should not store binaries).

tbot now copies all binaries found in "binariessubpath"
in tftp directory for the board to U-Boots build directory
on build host.

Signed-off-by: Heiko Schocher <hs@denx.de>