tbottest: move boardspecific.py file to same place as tbot.ini
move the place of boardspecifc.py and labcallbacks.py
to the subdirectory where tbot.ini file is found, as this
allows us in one tbotconfig folder to have more than one
lab and board configurations.

tbottest expects now this files, where the tbot.ini
file is found (from the path which is passed through
tbot flag inifile)

To be backward compatible, for a boardspecific.py file
is searched in tbotconfig path. If found, a warning is
printed and this one is used. This may gets removed
in later versions.

Signed-off-by: Heiko Schocher <hs@denx.de>